+++ packages/integrations/src/transaction.ts
@@ -1,5 +1,7 @@
 import { Event, EventProcessor, Hub, Integration, StackFrame } from '@sentry/types';
 
+import { getFramesFromEvent } from './helpers';
+
 /** Add node transaction to the event */
 export class Transaction implements Integration {
   /**
@@ -28,7 +30,7 @@ export class Transaction implements Integration {
    * @inheritDoc
    */
   public process(event: Event): Event {
-    const frames = this._getFramesFromEvent(event);
+    const frames = getFramesFromEvent(event) || [];
 
     // use for loop so we don't have to reverse whole frames array
     for (let i = frames.length - 1; i >= 0; i--) {
@@ -41,12 +43,6 @@ export class Transaction implements Integration {
     }
 
     return event;
-  }
-
-  /** JSDoc */
-  private _getFramesFromEvent(event: Event): StackFrame[] {
-    const exception = event.exception && event.exception.values && event.exception.values[0];
-    return (exception && exception.stacktrace && exception.stacktrace.frames) || [];
   }
 
   /** JSDoc */