+++ packages/node/src/backend.ts
@@ -130,9 +130,13 @@ export class NodeBackend extends BaseBackend<NodeOptions> {
       if (this._options.attachStacktrace && hint && hint.syntheticException) {
         const stack = hint.syntheticException ? extractStackFromError(hint.syntheticException) : [];
         parseStack(stack, this._options).then(frames => {
-          event.stacktrace = {
-            frames: prepareFramesForEvent(frames),
-          };
+          event.threads = [
+            {
+              stacktrace: {
+                frames: prepareFramesForEvent(frames),
+              },
+            },
+          ];
           resolve(event);
         });
       } else {